Move Mirror settings to dedicated tab
