+++ trace/Views/Settings/GeneralSettingsView.swift
@@ -6,7 +6,6 @@
 //
 
 import SwiftUI
-import AVFoundation
 import Carbon
 import UniformTypeIdentifiers
 import os.log
@@ -36,11 +35,8 @@ struct GeneralSettingsView: View {
     @ObservedObject private var permissionManager = PermissionManager.shared
     @State private var accessibilityEnabled = false
     @State private var calendarEnabled = false
-    @State private var cameraAuthorizationStatus: AVAuthorizationStatus = .notDetermined
     @State private var checkingPermissions = false
     @State private var requestingCalendarPermission = false
-    @State private var requestingCameraPermission = false
-    @State private var availableMirrorCameras: [AVCaptureDevice] = []
     let onLaunchAtLoginChange: (Bool) -> Void
     let onHotkeyRecord: (UInt32, UInt32) -> Void
     let onHotkeyReset: () -> Void
@@ -117,7 +113,9 @@ struct GeneralSettingsView: View {
                         openAccessibilitySettings()
                     }
                 )
-                
+
+                NativeSettingsDivider()
+
                 PermissionRow(
                     title: "Calendar Access",
                     subtitle: "Enables calendar event search",
@@ -133,23 +131,6 @@ struct GeneralSettingsView: View {
                     buttonTitle: calendarEnabled ? "Open Settings" : "Request Permission"
                 )
 
-                NativeSettingsDivider()
-
-                PermissionRow(
-                    title: "Camera Access",
-                    subtitle: "Enables the local Mirror preview",
-                    icon: "video",
-                    status: requestingCameraPermission || checkingPermissions ? .checking : cameraPermissionStatus,
-                    action: {
-                        if cameraAuthorizationStatus == .notDetermined {
-                            requestCameraPermission()
-                        } else {
-                            openCameraPrivacySettings()
-                        }
-                    },
-                    buttonTitle: cameraPermissionButtonTitle
-                )
-                
                 NativeSettingsDivider()
                 
                 HStack {
@@ -175,37 +156,6 @@ struct GeneralSettingsView: View {
                 .frame(minHeight: 44)
             } footer: {
                 Text("All permissions are optional. Features that need a permission may be unavailable until you grant it, and you can change access any time in System Settings.")
-            }
-
-            NativeSettingsSection("Mirror") {
-                NativeSettingsRow(
-                    title: "Camera",
-                    subtitle: mirrorCameraSubtitle
-                ) {
-                    Picker("", selection: Binding(
-                        get: { settingsManager.settings.mirrorCameraDeviceID },
-                        set: { settingsManager.updateMirrorCameraDeviceID($0) }
-                    )) {
-                        Text("System Default").tag("")
-
-                        ForEach(availableMirrorCameras, id: \.uniqueID) { device in
-                            Text(MirrorManager.displayName(for: device)).tag(device.uniqueID)
-                        }
-
-                        // Keep a previously saved camera visible if it is currently disconnected.
-                        if !settingsManager.settings.mirrorCameraDeviceID.isEmpty,
-                           !availableMirrorCameras.contains(where: {
-                               $0.uniqueID == settingsManager.settings.mirrorCameraDeviceID
-                           }) {
-                            Text("Unavailable camera").tag(settingsManager.settings.mirrorCameraDeviceID)
-                        }
-                    }
-                    .pickerStyle(.menu)
-                    .frame(minWidth: 200, maxWidth: 260)
-                    .disabled(cameraAuthorizationStatus != .authorized)
-                }
-            } footer: {
-                Text("Choose which camera Mirror uses. System Default follows the camera recommended by macOS. Continuity Camera (iPhone/iPad) appears when the device is nearby, unlocked, and camera access is granted.")
             }
 
             NativeSettingsSection("Interface") {
@@ -354,21 +304,9 @@ struct GeneralSettingsView: View {
             resultsLayout = ResultsLayout(rawValue: settingsManager.settings.resultsLayout) ?? .compact
             showMenuBarIcon = settingsManager.settings.showMenuBarIcon
             accentColor = settingsManager.selectedAccent
-            refreshAvailableMirrorCameras()
             
             // Check permissions
             checkPermissions()
-        }
-        .onReceive(NotificationCenter.default.publisher(for: AVCaptureDevice.wasConnectedNotification)) { _ in
-            refreshAvailableMirrorCameras()
-        }
-        .onReceive(NotificationCenter.default.publisher(for: AVCaptureDevice.wasDisconnectedNotification)) { _ in
-            refreshAvailableMirrorCameras()
-        }
-        .onChange(of: settingsManager.settings.mirrorCameraDeviceID) { _, _ in
-            Task { @MainActor in
-                ServiceContainer.shared.mirrorManager.applyCameraSelectionChange()
-            }
         }
         .onDisappear {
             stopRecording()
@@ -571,9 +509,6 @@ struct GeneralSettingsView: View {
         
         // Check calendar permissions
         checkCalendarPermissions()
-
-        // Check camera permissions
-        checkCameraPermissions()
     }
     
     private func checkAccessibilityPermissions() {
@@ -610,64 +545,6 @@ struct GeneralSettingsView: View {
         }
     }
 
-    private var cameraPermissionStatus: PermissionStatus {
-        switch cameraAuthorizationStatus {
-        case .authorized:
-            return .granted
-        case .notDetermined:
-            return .notDetermined
-        case .denied, .restricted:
-            return .denied
-        @unknown default:
-            return .denied
-        }
-    }
-
-    private var cameraPermissionButtonTitle: String {
-        cameraAuthorizationStatus == .notDetermined ? "Request Permission" : "Open Settings"
-    }
-
-    private var mirrorCameraSubtitle: String {
-        if cameraAuthorizationStatus != .authorized {
-            return "Grant camera access to choose a Mirror camera"
-        }
-        if availableMirrorCameras.isEmpty {
-            return "No cameras detected"
-        }
-        let cameraCount = availableMirrorCameras.count
-        let cameraLabel = cameraCount == 1 ? "camera" : "cameras"
-        let continuityCount = availableMirrorCameras.filter {
-            $0.isContinuityCamera || $0.deviceType == .continuityCamera
-        }.count
-        if continuityCount > 0 {
-            return "\(cameraCount) \(cameraLabel) · includes Continuity Camera"
-        }
-        return "\(cameraCount) \(cameraLabel) available for Mirror"
-    }
-
-    private func checkCameraPermissions() {
-        cameraAuthorizationStatus = AVCaptureDevice.authorizationStatus(for: .video)
-        refreshAvailableMirrorCameras()
-    }
-
-    private func refreshAvailableMirrorCameras() {
-        availableMirrorCameras = MirrorManager.availableCameras()
-    }
-
-    private func requestCameraPermission() {
-        guard !requestingCameraPermission else { return }
-
-        Task {
-            requestingCameraPermission = true
-            let granted = await AVCaptureDevice.requestAccess(for: .video)
-            await MainActor.run {
-                cameraAuthorizationStatus = granted ? .authorized : AVCaptureDevice.authorizationStatus(for: .video)
-                requestingCameraPermission = false
-                refreshAvailableMirrorCameras()
-            }
-        }
-    }
-    
     /// Requests calendar permissions from the user and enables calendar search if granted
     private func requestCalendarPermission() {
         guard !requestingCalendarPermission else { return }
@@ -704,11 +581,6 @@ struct GeneralSettingsView: View {
         }
     }
 
-    private func openCameraPrivacySettings() {
-        if let url = URL(string: "x-apple.systempreferences:com.apple.preference.security?Privacy_Camera") {
-            NSWorkspace.shared.open(url)
-        }
-    }
 }
 
 // MARK: - Permission Types
